+++ x86-sse2.h
@@ -27,47 +27,48 @@
 #include <stdint.h>
 
 #ifdef __cplusplus
-extern "C" {
+extern "C"
+{
 #endif
 
-void aligned_block_copy_movsb(int64_t * __restrict dst,
-                              int64_t * __restrict src,
-                              int                  size);
-void aligned_block_copy_movsd(int64_t * __restrict dst,
-                              int64_t * __restrict src,
-                              int                  size);
-void aligned_block_fill_stosb(int64_t * __restrict dst,
-                              int64_t * __restrict src,
-                              int                  size);
+    void aligned_block_copy_movsb(int64_t *__restrict dst,
+                                  int64_t *__restrict src,
+                                  size_t size);
+    void aligned_block_copy_movsd(int64_t *__restrict dst,
+                                  int64_t *__restrict src,
+                                  size_t size);
+    void aligned_block_fill_stosb(int64_t *__restrict dst,
+                                  int64_t *__restrict src,
+                                  size_t size);
 
-void aligned_block_copy_sse2(int64_t * __restrict dst,
-                             int64_t * __restrict src,
-                             int                  size);
-void aligned_block_copy_nt_sse2(int64_t * __restrict dst,
-                                int64_t * __restrict src,
-                                int                  size);
+    void aligned_block_copy_sse2(int64_t *__restrict dst,
+                                 int64_t *__restrict src,
+                                 size_t size);
+    void aligned_block_copy_nt_sse2(int64_t *__restrict dst,
+                                    int64_t *__restrict src,
+                                    size_t size);
 
-void aligned_block_copy_pf32_sse2(int64_t * __restrict dst,
-                                  int64_t * __restrict src,
-                                  int                  size);
-void aligned_block_copy_pf64_sse2(int64_t * __restrict dst,
-                                  int64_t * __restrict src,
-                                  int                  size);
+    void aligned_block_copy_pf32_sse2(int64_t *__restrict dst,
+                                      int64_t *__restrict src,
+                                      size_t size);
+    void aligned_block_copy_pf64_sse2(int64_t *__restrict dst,
+                                      int64_t *__restrict src,
+                                      size_t size);
 
-void aligned_block_copy_nt_pf32_sse2(int64_t * __restrict dst,
-                                     int64_t * __restrict src,
-                                     int                  size);
-void aligned_block_copy_nt_pf64_sse2(int64_t * __restrict dst,
-                                     int64_t * __restrict src,
-                                     int                  size);
+    void aligned_block_copy_nt_pf32_sse2(int64_t *__restrict dst,
+                                         int64_t *__restrict src,
+                                         size_t size);
+    void aligned_block_copy_nt_pf64_sse2(int64_t *__restrict dst,
+                                         int64_t *__restrict src,
+                                         size_t size);
 
-void aligned_block_fill_sse2(int64_t * __restrict dst,
-                             int64_t * __restrict src,
-                             int                  size);
+    void aligned_block_fill_sse2(int64_t *__restrict dst,
+                                 int64_t *__restrict src,
+                                 size_t size);
 
-void aligned_block_fill_nt_sse2(int64_t * __restrict dst,
-                                int64_t * __restrict src,
-                                int                  size);
+    void aligned_block_fill_nt_sse2(int64_t *__restrict dst,
+                                    int64_t *__restrict src,
+                                    size_t size);
 
 #ifdef __cplusplus
 }
